Upgrade React SDK, refactor useListConversations
