+++ hooks/useSendMessage.ts
@@ -7,7 +7,7 @@ import {
   useSendMessage as useSendMessageHook,
   useStartConversation,
 } from "@xmtp/react-sdk";
-import { Conversation } from "@xmtp/xmtp-js";
+import { Conversation } from "@xmtp/react-sdk";
 
 const useSendMessage = (conversationId: address) => {
   const conversations = useXmtpStore((state) => state.conversations);
