+++ hooks/useStartFeedbackConvo.ts
@@ -1,4 +1,4 @@
-import { DecodedMessage } from "@xmtp/xmtp-js";
+import { DecodedMessage } from "@xmtp/react-sdk";
 import { useEffect } from "react";
 import { XMTP_FEEDBACK_ADDRESS, getConversationId } from "../helpers";
 import { useXmtpStore } from "../store/xmtp";
