+++ package.json
@@ -33,8 +33,7 @@
     "@heroicons/react": "^1.0.5",
     "@rainbow-me/rainbowkit": "^0.12.8",
     "@tailwindcss/line-clamp": "^0.4.2",
-    "@xmtp/react-sdk": "^1.0.0-preview.25",
-    "@xmtp/xmtp-js": "^8.1.0",
+    "@xmtp/react-sdk": "^1.0.0-preview.27",
     "date-fns": "^2.29.3",
     "emojibase": "^6.1.0",
     "ethers": "^5.7.2",
